ci: cleanup develop
