+++ __tests__/utils/ellipticalCurve.test.ts
@@ -1,36 +1,85 @@
-import { ec, hash } from '../../src';
+import { ec } from '../../src';
+import { StarknetChainId } from '../../src/constants';
+import {
+  calculateTransactionHash,
+  computeHashOnElements,
+  transactionVersion,
+} from '../../src/utils/hash';
+import { fromCallsToExecuteCalldataWithNonce } from '../../src/utils/transaction';
 
-test('getStarkKey()', () => {
+test('getKeyPair()', () => {
   const privateKey = '0x019800ea6a9a73f94aee6a3d2edf018fc770443e90c7ba121e8303ec6b349279';
   const starkKey = ec.starkCurve.getStarkKey(privateKey);
+  // somehow needed, returns error else
   expect(starkKey).toBe('0x33f45f07e1bd1a51b45fc24ec8c8c9908db9e42191be9e169bfcac0c0d99745');
 });
 
-test('getPublicKey()', () => {
-  const privateKey = '0x019800ea6a9a73f94aee6a3d2edf018fc770443e90c7ba121e8303ec6b349279';
-  const fullPubKey = ec.starkCurve.getPublicKey(privateKey);
-  // somehow needed, returns error else
-  expect(fullPubKey).toBe(
-    '0x04033f45f07e1bd1a51b45fc24ec8c8c9908db9e42191be9e169bfcac0c0d997450319d0f53f6ca077c4fa5207819144a2a4165daef6ee47a7c1d06c0dcaa3e456'
+test('pedersen()', () => {
+  const own = ec.starkCurve.pedersen('0x12773', '0x872362');
+  expect(own).toMatchSnapshot();
+});
+
+test('pedersen() with 0', () => {
+  const own = ec.starkCurve.pedersen('0x12773', '0x0');
+  expect(own).toMatchSnapshot();
+});
+
+test('computeHashOnElements()', () => {
+  const array = ['1', '2', '3', '4'];
+  expect(computeHashOnElements(array)).toBe(
+    '0x66bd4335902683054d08a0572747ea78ebd9e531536fb43125424ca9f902084'
+  );
+  expect(array).toStrictEqual(['1', '2', '3', '4']);
+
+  expect(computeHashOnElements(['1', '2'])).toBe(
+    '0x501a3a8e6cd4f5241c639c74052aaa34557aafa84dd4ba983d6443c590ab7df'
   );
 });
 
-test('verify signed message with full public key()', () => {
-  const privk = '0x019800ea6a9a73f94aee6a3d2edf018fc770443e90c7ba121e8303ec6b349279';
-  const account = '0x33f45f07e1bd1a51b45fc24ec8c8c9908db9e42191be9e169bfcac0c0d99745';
-  const price = '1';
-  const hashMsg = hash.pedersen(account, price);
-  const signature = ec.starkCurve.sign(hashMsg, privk);
-  const pubKey = ec.starkCurve.getPublicKey(privk);
-  expect(ec.starkCurve.verify(signature, hashMsg, pubKey)).toBe(true);
+test('hashMessage()', () => {
+  const privateKey = '0x019800ea6a9a73f94aee6a3d2edf018fc770443e90c7ba121e8303ec6b349279';
+  const account = '2007067565103695475819120104515800035851923905855118399071773059478896040938';
+  const transactions = [
+    {
+      contractAddress:
+        '3290661298119599979891444342541795905081168856323302956721669397616389152866',
+      entrypoint: 'set_number',
+      calldata: ['47'],
+    },
+  ];
+  const nonce = '3';
+  const maxFee = '0';
+  const calldata = fromCallsToExecuteCalldataWithNonce(transactions, nonce);
+
+  const hashMsg = calculateTransactionHash(
+    account,
+    transactionVersion,
+    calldata,
+    maxFee,
+    StarknetChainId.SN_GOERLI,
+    nonce
+  );
+
+  expect(hashMsg).toMatchInlineSnapshot(
+    `"0x6d1706bd3d1ba7c517be2a2a335996f63d4738e2f182144d078a1dd9997062e"`
+  );
+
+  const { r, s } = ec.starkCurve.sign(hashMsg, privateKey);
+
+  expect(r.toString()).toMatchInlineSnapshot(
+    `"1427981024487605678086498726488552139932400435436186597196374630267616399345"`
+  );
+  expect(s.toString()).toMatchInlineSnapshot(
+    `"1853664302719670721837677288395394946745467311923401353018029119631574115563"`
+  );
 });
 
-test('verify signed message with Starknet public Key()', () => {
-  const privk = '0x019800ea6a9a73f94aee6a3d2edf018fc770443e90c7ba121e8303ec6b349279';
+test('verify signed message()', () => {
+  const pk = '0x019800ea6a9a73f94aee6a3d2edf018fc770443e90c7ba121e8303ec6b349279';
   const account = '0x33f45f07e1bd1a51b45fc24ec8c8c9908db9e42191be9e169bfcac0c0d99745';
   const price = '1';
-  const hashMsg = hash.pedersen(account, price);
-  const signature = ec.starkCurve.sign(hashMsg, privk);
-  const pubKey = ec.starkCurve.getStarkKey(privk);
-  expect(ec.starkCurve.verify(signature, hashMsg, pubKey)).toBe(true);
+  const hashMsg = ec.starkCurve.pedersen(account, price);
+  const signature = ec.starkCurve.sign(hashMsg, pk);
+  const pubKey = ec.starkCurve.getPublicKey(pk);
+  expect(ec.starkCurve.verify(signature.toDERHex(), hashMsg, pubKey)).toBe(true);
 });
