+++ www/docs/guides/signature.md
@@ -6,9 +6,6 @@ sidebar_position: 14
 
 You can use Starknet.js to sign a message outside of the network, using the standard methods of hash and sign of Starknet. In this way, in some cases, you can avoid paying fees to store data in-chain; you transfer the signed message off-chain, and the recipient can verify (without fee) on-chain the validity of the message.
 
-> From Starknet.js V5.0, a new and way more faster crypto library is available. Nevertheless, the API is significantly modified.  
-> The crypto V4 API are still available, for ascending compatibility purpose, but do not use them anymore for your new code.
-
 ## Sign and send a message
 
 Your message has to be an array of `BigNumberish`. First calculate the hash of this message, then calculate the signature.
@@ -18,22 +15,22 @@ Your message has to be an array of `BigNumberish`. First calculate the hash of t
 ```typescript
 import {ec, hash, num, json, Contract } from "starknet";
 
-const privateKey = '0x5b7d4f8710b3581ebb2b8b74efaa23d25ab0ffea2a4f3e269bf91bf9f63d633';
-const starknetPubKey = ec.starkCurve.getStarkKey(privateKey);
+const privateKey = "0x1234567890987654321";
+const starkKeyPair = ec.getKeyPair(privateKey);
+const starknetPublicKey = ec.getStarkKey(starkKeyPair);
+const fullPublicKey=starkKeyPair.getPublic("hex");
 
-const message : number.BigNumberish[] = [1, 128, 18, 14];
+const message : BigNumberish[] = [1, 128, 18, 14];
 
 const msgHash = hash.computeHashOnElements(message);
-const signature = ec.starkCurve.sign(msgHash, privateKey);
+const signature = ec.sign(starkKeyPair, msgHash);
 ```
 
 Then you can send, by any means, to the recipient of the message:
 
 - the message.
 - the signature.
-- the Starknet public key, or the address of the wallet related to this private key.
-
-> You can perform a multi-signature of a message. You have just to create an array of (signatures + public keys), and then to verify all these pairs.
+- the full public key (or a wallet address).
 
 ## Receive and verify a message
 
@@ -44,22 +41,44 @@ On receiver side, you can verify that:
 
 2 ways to perform this verification:
 
-- off-chain, using the Starknet public key (very fast, but only for standard Starknet hash & sign).
+- off-chain, using the full public key (very fast, but only for standard Starknet hash & sign).
 - on-chain, using the account address (slow, add workload to the node/sequencer, but can manage exotic account abstraction about hash or sign).
 
 ### Verify outside of Starknet:
 
-The sender provides the message, the signature and the Starknet public key. Verification:
+The sender provides the message, the signature and the full public key. Verification:
 
 ```typescript
+const starkKeyPair1 = ec.getKeyPairFromPublicKey(fullPublicKey);
 const msgHash1 = hash.computeHashOnElements(message);
-const result1 = ec.starkCurve.verify(signature, msgHash, starknetPubKey);
+const result1 = ec.verify(starkKeyPair1, msgHash1, signature);
 console.log("Result (boolean) =", result1);
+```
+
+> The sender can also provide their account address. Then you can check that this full public key is linked to this account. The pubKey that you can read in the account contract is part (part X) of the full pubKey (parts X & Y):
+
+Read the pubKey of the account :
+
+```typescript
+const provider = new Provider({ sequencer: { baseUrl: "http://127.0.0.1:5050" } }); //devnet
+const compiledAccount = json.parse(fs.readFileSync("./compiled_contracts/Account_0_5_1.json").toString("ascii"));
+const accountAddress ="0x...."; // account of sender
+const contractAccount = new Contract(compiledAccount.abi, accountAddress, provider);
+const pubKey3 = await contractAccount.call("getPublicKey");
+```
+
+Check that the pubKey of the account is part of the full pubKey:
+
+```typescript
+const isFullPubKeyRelatedToAccount: boolean =
+    BigInt(pubKey3.publicKey.toString()) ==
+    BigInt(encode.addHexPrefix(fullPublicKey.slice(4, 68)));
+console.log("Result (boolean)=", isFullPubKeyRelatedToAccount);
 ```
 
 ### Verify in Starknet network, with the account:
 
-The sender can provide an account address, in spite of a Starknet public key.
+The sender can provide an account address, in spite of a full public key.
 
 ```typescript
 const provider = new Provider({ sequencer: { baseUrl: "http://127.0.0.1:5050" } }); //devnet
@@ -166,7 +185,7 @@ const msgHash5 = typedData.getMessageHash(typedDataValidate, accountAddress);
 // The call of isValidSignature will generate an error if not valid
 let result5: boolean;
 try {
-    await contractAccount.call("isValidSignature", [msgHash5, signature4]);
+    await contractAccount.call("isValidSignature", [msgHash5, signature5]);
     result5 = true;
 } catch {
     result5 = false;
